+++ CHANGELOG.md
@@ -3,6 +3,11 @@
 ## [v1.14.11 - 2024-11-08]
 ### New
 - common: add ReadStdinToString and ReadStdinByLine functions and tests
+- dblib: add ssl tns 
+### Changed
+- dblib: modify tns ldap tests
+- update dependencies
+
 ## [v1.14.10 - 2024-09-04]
 ### Changed
 - update dns tests